requirements.txt
requests>=2.31
urllib3>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
apache-airflow>=2.8
